Using int64 instead of int32 for file descriptors to make sure we won't get OverflowException during an attempt to convert DangerousGetHandle to a file descriptor.

+++ src/Lucene.Net/DNX/FileStreamExtensions.cs
@@ -21,7 +21,7 @@ namespace Lucene.Net.DNX
                 if (position != 0 && length != stream.Length)
                     throw new InvalidOperationException("On Posix platform we allow to lock only an entire file"); // we have only such usage in Lucene.NET
 
-                if (PosixNativeFileMethods.flock(stream.SafeFileHandle.DangerousGetHandle().ToInt32(), PosixFileLockOptions.LOCK_EX) != 0) // TODO arek how to get file descriptor?
+                if (PosixNativeFileMethods.flock(stream.SafeFileHandle.DangerousGetHandle().ToInt64(), PosixFileLockOptions.LOCK_EX) != 0)
                     throw new InvalidOperationException($"Failure when trying to lock file in posix, error code was: {Marshal.GetLastWin32Error()}");
             }
             else
@@ -45,13 +45,12 @@ namespace Lucene.Net.DNX
             if (stream.SafeFileHandle.IsClosed)
                 throw new InvalidOperationException("File is closed");
 
-
             if (PlatformDetection.RunningOnPosix)
             {
                 if (position != 0 && length != stream.Length)
                     throw new InvalidOperationException("On Posix platform we allow to lock only an entire file"); // we have only such usage in Lucene.NET
 
-                if (PosixNativeFileMethods.flock(stream.SafeFileHandle.DangerousGetHandle().ToInt32(), PosixFileLockOptions.LOCK_UN) != 0) // TODO arek how to get file descriptor?
+                if (PosixNativeFileMethods.flock(stream.SafeFileHandle.DangerousGetHandle().ToInt64(), PosixFileLockOptions.LOCK_UN) != 0)
                     throw new InvalidOperationException($"Failure when trying to unlock file in posix, error code was: {Marshal.GetLastWin32Error()}");
             }
             else
